requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
black==25.12.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.5
daphne==4.2.1
defusedxml==0.7.1
Django==5.1.3
django-allauth==65.10.0
django-bootstrap5==24.3
django-celery-beat==2.8.1
django-cleanup==9.0.0
django-debug-toolbar==6.0.0
django-decouple==2.1
django-environ==0.12.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-redis==6.0.0
django-timezone-field==7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
flake8==7.3.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
Markdown==3.7
marshmallow==3.23.1
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
nltk==3.9.1
oauthlib==3.3.1
packaging==24.2
pathspec==1.0.0
platformdirs==4.5.1
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.4.2
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytokens==0.3.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sqlparse==0.5.2
text-unidecode==1.3
textblob==0.19.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
